# Optimize scalar and avx2 implementations