--- a/_gen/go.mod
+++ b/_gen/go.mod
@@ -1,5 +1,5 @@
 module github.com/minio/md5-simd/_gen
 
-go 1.14
+go 1.16
 
-require github.com/mmcloughlin/avo v0.0.0-20210104032911-599bdd1269f4 // indirect
+require github.com/mmcloughlin/avo v0.2.0