Use https://github.com/animetosho/md5-optimisation#dependency-shortcut-in-g-function and https://github.com/animetosho/md5-optimisation#h-function-re-use for shorter dependency chain.

Cleanup in AVX2 removing superfluous loads/moves.

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkAvx2/32KB-32                  201378        194863        -3.24%
BenchmarkAvx2/64KB-32                  321507        303803        -5.51%
BenchmarkAvx2/128KB-32                 594137        577175        -2.85%
BenchmarkAvx2/256KB-32                 1089630       1014160       -6.93%
BenchmarkAvx2/512KB-32                 2077582       1959077       -5.70%
BenchmarkAvx2/1MB-32                   4191188       3949610       -5.76%
BenchmarkAvx2/2MB-32                   8439181       8042106       -4.71%
BenchmarkAvx2/4MB-32                   16655067      15739187      -5.50%
BenchmarkAvx2/8MB-32                   33017781      31620932      -4.23%
BenchmarkAvx2SingleWriter/32KB-32      41765         39763         -4.79%
BenchmarkAvx2SingleWriter/64KB-32      81884         76866         -6.13%
BenchmarkAvx2SingleWriter/128KB-32     166802        155819        -6.58%
BenchmarkAvx2SingleWriter/256KB-32     329145        306292        -6.94%
BenchmarkAvx2SingleWriter/512KB-32     653422        616564        -5.64%
BenchmarkAvx2SingleWriter/1MB-32       1303555       1237368       -5.08%
BenchmarkAvx2SingleWriter/2MB-32       2596346       2441836       -5.95%
BenchmarkAvx2SingleWriter/4MB-32       5151380       4885766       -5.16%
BenchmarkAvx2SingleWriter/8MB-32       10324461      9765875       -5.41%
```